+++ football365/migrations/0002_auto__add_field_call_client_id.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+import datetime
+from south.db import db
+from south.v2 import SchemaMigration
+from django.db import models
+
+
+class Migration(SchemaMigration):
+
+    def forwards(self, orm):
+        # Adding field 'Call.client_id'
+        db.add_column('football365_call', 'client_id',
+                      self.gf('django.db.models.fields.CharField')(max_length=32, null=True, blank=True),
+                      keep_default=False)
+
+
+    def backwards(self, orm):
+        # Deleting field 'Call.client_id'
+        db.delete_column('football365_call', 'client_id')
+
+
+    models = {
+        'football365.call': {
+            'Meta': {'ordering': "('title',)", 'object_name': 'Call'},
+            'call_type': ('django.db.models.fields.CharField', [], {'max_length': '32'}),
+            'client_id': ('django.db.models.fields.CharField', [], {'max_length': '32', 'null': 'True', 'blank': 'True'}),
+            'football365_service_id': ('django.db.models.fields.PositiveIntegerField', [], {}),
+            'id': ('django.db.models.fields.AutoField', [], {'primary_key': 'True'}),
+            'title': ('django.db.models.fields.CharField', [], {'max_length': '256'})
+        }
+    }
+
+    complete_apps = ['football365']