+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 =========
 
+next
+----
+#. Convert to south.
+#. Allow client account id to be overridden per call. Needed because Teamtalk changed the rules.
+
 0.0.3
 -----
 #. Handle XML feed change.